Add RestoreGroup method for IDataStorage
RestoreGroup returns an IEnumerable of all objects stored in a group.

A requirement is for all objects in a group to be of same type. If there is at least one object that doesn't cast well into the provided type, the whole method returns an empty list. (could be improved later)